Перенесла по папкам файлы
