+++ app/src/main/AndroidManifest.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools">
     <uses-permission android:name="android.permission.INTERNET"/>
     <application
-        android:name=".App"
+        android:name=".ui.App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -13,18 +13,18 @@
         android:theme="@style/Theme.MaterialComponents.DayNight.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".activity.SearchActivity"
+            android:name=".ui.SearchActivity"
             android:exported="false" />
         <activity
-            android:name=".activity.MediaActivity"
-            android:exported="false" />
-
-        <activity
-            android:name=".activity.MediaPlayerActivity"
+            android:name=".ui.MediaActivity"
             android:exported="false" />
 
         <activity
-            android:name=".activity.SettingsActivity"
+            android:name=".ui.MediaPlayerActivity"
+            android:exported="false" />
+
+        <activity
+            android:name=".ui.SettingsActivity"
             android:exported="true">
 
         </activity>
@@ -34,7 +34,7 @@
             android:resource="@xml/actions" />
 
         <activity
-            android:name=".activity.MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
 
             <intent-filter>
