--- app/src/main/java/com/example/playlistmaker/OnTrackClickListener.kt
@@ -1,7 +1,0 @@
-package com.example.playlistmaker
-
-import com.example.playlistmaker.data.Track
-
-interface OnTrackClickListener {
-    fun onTrackClick(track: Track)
-}
