--- app/src/main/java/com/example/playlistmaker/SearchHistory.kt
@@ -1,11 +1,0 @@
-package com.example.playlistmaker
-
-import com.example.playlistmaker.data.Track
-
-interface SearchTrackHistory {
-
-    fun addTrack(track: Track)
-
-    fun getTracks(): List<Track>
-
-}
