+++ app/src/main/java/com/example/playlistmaker/data/dto/ITunesResponse.kt
@@ -1,0 +1,6 @@
+package com.example.playlistmaker.data.dto
+
+class ITunesResponse(
+    val resultCount: Int,
+    val results: List<Track>,
+)
