+++ app/src/main/java/com/example/playlistmaker/data/dto/TrackTime.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data
+package com.example.playlistmaker.data.dto
 
 import java.io.Serializable
