--- app/src/main/java/com/example/playlistmaker/network/ITunesResponse.kt
@@ -1,8 +1,0 @@
-package com.example.playlistmaker.network
-
-import com.example.playlistmaker.data.Track
-
-class ITunesResponse(
-    val resultCount: Int,
-    val results: List<Track>,
-)
