+++ app/src/main/java/com/example/playlistmaker/ui/MediaActivity.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.activity
+package com.example.playlistmaker.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
