+++ app/src/main/java/com/example/playlistmaker/ui/SearchHistory.kt
@@ -1,0 +1,11 @@
+package com.example.playlistmaker.ui
+
+import com.example.playlistmaker.data.dto.Track
+
+interface SearchTrackHistoryRepository {
+
+    fun addTrack(track: Track)
+
+    fun getTracks(): List<Track>
+
+}
