+++ app/src/main/java/com/example/playlistmaker/ui/recyclerview/OnTrackClickListener.kt
@@ -1,0 +1,7 @@
+package com.example.playlistmaker.ui.recyclerview
+
+import com.example.playlistmaker.data.dto.Track
+
+interface OnTrackClickListener {
+    fun onTrackClick(track: Track)
+}
