+++ app/src/main/java/com/example/playlistmaker/ui/recyclerview/TrackAdapter.kt
@@ -1,12 +1,11 @@
-package com.example.playlistmaker.recyclerview
+package com.example.playlistmaker.ui.recyclerview
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 
 import androidx.recyclerview.widget.RecyclerView
-import com.example.playlistmaker.OnTrackClickListener
 import com.example.playlistmaker.R
-import com.example.playlistmaker.data.Track
+import com.example.playlistmaker.data.dto.Track
 
 class TrackAdapter(
     private val data: List<Track>,
