+++ app/src/main/java/com/example/playlistmaker/ui/recyclerview/TrackViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.recyclerview
+package com.example.playlistmaker.ui.recyclerview
 
 import android.view.View
 import android.widget.ImageView
@@ -7,7 +7,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.example.playlistmaker.R
-import com.example.playlistmaker.data.Track
+import com.example.playlistmaker.data.dto.Track
 
 class TrackViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
